bug: fix disposing controller
