+++ example/lib/barcode_scanner_controller.dart
@@ -1,0 +1,114 @@
+import 'package:flutter/material.dart';
+import 'package:mobile_scanner/mobile_scanner.dart';
+
+class BarcodeScannerWithController extends StatefulWidget {
+  const BarcodeScannerWithController({Key? key}) : super(key: key);
+
+  @override
+  _BarcodeScannerWithControllerState createState() =>
+      _BarcodeScannerWithControllerState();
+}
+
+class _BarcodeScannerWithControllerState
+    extends State<BarcodeScannerWithController>
+    with SingleTickerProviderStateMixin {
+  String? barcode;
+
+  MobileScannerController controller = MobileScannerController(
+    torchEnabled: true,
+    // facing: CameraFacing.front,
+  );
+
+  @override
+  Widget build(BuildContext context) {
+    return MaterialApp(
+      home: Scaffold(
+        backgroundColor: Colors.black,
+        body: Builder(builder: (context) {
+          return Stack(
+            children: [
+              MobileScanner(
+                  controller: controller,
+                  fit: BoxFit.contain,
+                  // controller: MobileScannerController(
+                  //   torchEnabled: true,
+                  //   facing: CameraFacing.front,
+                  // ),
+                  onDetect: (barcode, args) {
+                    if (this.barcode != barcode.rawValue) {
+                      setState(() {
+                        this.barcode = barcode.rawValue;
+                      });
+                    }
+                  }),
+              Align(
+                alignment: Alignment.bottomCenter,
+                child: Container(
+                  alignment: Alignment.bottomCenter,
+                  height: 100,
+                  color: Colors.black.withOpacity(0.4),
+                  child: Row(
+                    crossAxisAlignment: CrossAxisAlignment.center,
+                    mainAxisAlignment: MainAxisAlignment.spaceEvenly,
+                    children: [
+                      IconButton(
+                        color: Colors.white,
+                        icon: ValueListenableBuilder(
+                          valueListenable: controller.torchState,
+                          builder: (context, state, child) {
+                            switch (state as TorchState) {
+                              case TorchState.off:
+                                return const Icon(Icons.flash_off,
+                                    color: Colors.grey);
+                              case TorchState.on:
+                                return const Icon(Icons.flash_on,
+                                    color: Colors.yellow);
+                            }
+                          },
+                        ),
+                        iconSize: 32.0,
+                        onPressed: () => controller.toggleTorch(),
+                      ),
+                      Center(
+                        child: SizedBox(
+                          width: MediaQuery.of(context).size.width - 120,
+                          height: 50,
+                          child: FittedBox(
+                            child: Text(
+                              barcode ?? 'Scan something!',
+                              overflow: TextOverflow.fade,
+                              style: Theme.of(context)
+                                  .textTheme
+                                  .headline4!
+                                  .copyWith(color: Colors.white),
+                            ),
+                          ),
+                        ),
+                      ),
+                      IconButton(
+                        color: Colors.white,
+                        icon: ValueListenableBuilder(
+                          valueListenable: controller.cameraFacingState,
+                          builder: (context, state, child) {
+                            switch (state as CameraFacing) {
+                              case CameraFacing.front:
+                                return const Icon(Icons.camera_front);
+                              case CameraFacing.back:
+                                return const Icon(Icons.camera_rear);
+                            }
+                          },
+                        ),
+                        iconSize: 32.0,
+                        onPressed: () => controller.switchCamera(),
+                      ),
+                    ],
+                  ),
+                ),
+              ),
+            ],
+          );
+        }),
+      ),
+    );
+  }
+}
